Avoid defining singleton construction in header file
That way, the call to new is always executed by code inside libastra.
This avoids the situation where a singleton gets created by a copy
of the constructor linked into an object file outside of libastra, such
as a .mex file, which would then also cause the vtable to be outside of
libastra. This situation would cause issues when .mex files are unloaded.