Fix rounding error in ParStripProjector2D::getProjectionWeightsCount
In the worst case this would lead to (nearly) empty storage for getMatrix(),
resulting in (near) explicit projection matrices. (These are only used for
exporting explicit sparse projection matrices to matlab/python; not for
FP/BP/reconstruction.)

This is a quick fix; ideally the affected code would use dynamic storage.