Improve conda dependency compatibility with conda-forge and nvidia channels
conda-forge now has a libastra package, and to limit the risk of API/ABI
incompatibility, we want to try to ensure our official astra-toolbox package
depends on our official libastra package, without accidentally accepting the
conda-forge package. The conda-forge package (currently) has a differently
formatted build string, so we make our libastra dependency more specific to
catch that.

There are a few cudatoolkit packages on the main and nvidia channels that
are non-chronologically numbered. Specifically:

11.0.221 is older than 11.0.3.
It has the same libcudart.so, so we accept >=11.0.3,<11.1.0a0 .

11.1.74 is older than 11.1.1.
It has the same libcudart.so, so we accept >=11.1.1,<11.2.0a0 .

11.2.72 is older than 11.2.1.
It has an older libcudart.so than 11.2.2, so we accept >=11.2.2,<11.2.72 .