Replace shared_ptr<CPart> in CGM
Split off internal uses of SJob to a new SJobInternal. The new struct
SJobInternal lacks a pointer to the output Part to clarify ownership of
Part.

In the *Internal family of structs, output
Parts are owned by the TJobSetInternal set, while input Parts are owned
by the SJobInternal jobs.

The outward-facing SJob struct owns both its input and output Parts.

This allows us to get rid of the shared_ptr-based memory management.