Add initial, very experimental support for AMD HIP
This allows running ASTRA on AMD GPUs.

This uses AMD's HIP compatibility layer with our existing CUDA code.
We include a wrapper header instead of cuda.h/cufft.h, to switch between
building with CUDA and building with HIP.

A notable missing feature is DLPack support for ROCm arrays.

Additionally, it is currently not possible to build both CUDA and HIP
support into a single binary. Moving the HIP-translated functions to a
separate namespace will later enable that.

From the user's side, it will currently appear as if CUDA is being used,
through for example, the CCudaProjector class. The required interface
changes for this are to be determined.